+++ src/frontend/js/components/SaleTunnel/components/SaleTunnelStepResume/index.tsx
@@ -1,10 +1,11 @@
 import { Button } from '@openfun/cunningham-react';
 import { defineMessages, FormattedMessage, useIntl } from 'react-intl';
+import { generatePath } from 'react-router-dom';
 import { SuccessIcon } from 'components/SuccessIcon';
-import { LearnerDashboardPaths } from 'widgets/Dashboard/utils/learnerRouteMessages';
 import { getDashboardBasename } from 'widgets/Dashboard/hooks/useDashboardRouter/getDashboardBasename';
 import { useSaleTunnelContext } from 'components/SaleTunnel/context';
-import { getDashboardRoutePath } from 'widgets/Dashboard/utils/dashboardRoutes';
+
+import { LearnerDashboardPaths } from 'widgets/Dashboard/utils/learnerRoutesPaths';
 
 const messages = defineMessages({
   congratulations: {
@@ -72,7 +73,7 @@ export const SaleTunnelStepResume = ({ next }: SaleTunnelStepResumeProps) => {
           <Button
             href={
               getDashboardBasename(intl.locale) +
-              getDashboardRoutePath(intl)(LearnerDashboardPaths.ORDER, { orderId: order!.id })
+              generatePath(LearnerDashboardPaths.ORDER, { orderId: order!.id })
             }
           >
             <FormattedMessage {...messages.ctaSignature} />
