+++ src/frontend/js/pages/DashboardAddressesManagement/DashboardCreateAddressLoader.tsx
@@ -1,5 +1,5 @@
-import { LearnerDashboardPaths } from 'widgets/Dashboard/utils/learnerRouteMessages';
 import { useDashboardNavigate } from 'widgets/Dashboard/hooks/useDashboardRouter';
+import { LearnerDashboardPaths } from 'widgets/Dashboard/utils/learnerRoutesPaths';
 import { DashboardCreateAddress } from './DashboardCreateAddress';
 
 /**
