+++ src/frontend/js/pages/DashboardContracts/index.spec.tsx
@@ -13,8 +13,9 @@ import { Deferred } from 'utils/test/deferred';
 import { History, HistoryContext } from 'hooks/useHistory';
 import { SessionProvider } from 'contexts/SessionContext';
 import { DashboardTest } from 'widgets/Dashboard/components/DashboardTest';
-import { LearnerDashboardPaths } from 'widgets/Dashboard/utils/learnerRouteMessages';
 import { ContractFactory } from 'utils/test/factories/joanie';
+
+import { LearnerDashboardPaths } from 'widgets/Dashboard/utils/learnerRoutesPaths';
 
 jest.mock('utils/context', () => ({
   __esModule: true,
