+++ src/frontend/js/pages/DashboardOrderLayout/index.tsx
@@ -1,17 +1,15 @@
-import { Outlet, useParams } from 'react-router-dom';
+import { generatePath, Outlet, useParams } from 'react-router-dom';
 import { useIntl } from 'react-intl';
 import { useMemo } from 'react';
-import {
-  getDashboardRouteLabel,
-  getDashboardRoutePath,
-} from 'widgets/Dashboard/utils/dashboardRoutes';
-import { LearnerDashboardPaths } from 'widgets/Dashboard/utils/learnerRouteMessages';
+import { getDashboardRouteLabel } from 'widgets/Dashboard/utils/dashboardRoutes';
 import { DashboardLayout } from 'widgets/Dashboard/components/DashboardLayout';
 import { useOmniscientOrder } from 'hooks/useOrders';
 import { useBreadcrumbsPlaceholders } from 'hooks/useBreadcrumbsPlaceholders';
 import { CourseLight, Product } from 'types/Joanie';
 import { LearnerDashboardSidebar } from 'widgets/Dashboard/components/LearnerDashboardSidebar';
 import { useCourseProduct } from 'hooks/useCourseProducts';
+
+import { LearnerDashboardPaths } from 'widgets/Dashboard/utils/learnerRoutesPaths';
 
 export const DashboardOrderLayout = () => {
   const params = useParams<{ orderId: string }>();
@@ -23,13 +21,12 @@ export const DashboardOrderLayout = () => {
   });
   const product = courseProduct?.item?.product;
   const intl = useIntl();
-  const getRoutePath = getDashboardRoutePath(intl);
   const getRouteLabel = getDashboardRouteLabel(intl);
 
   const links = useMemo(
     () => [
       {
-        to: getRoutePath(LearnerDashboardPaths.ORDER, { orderId: params.orderId }),
+        to: generatePath(LearnerDashboardPaths.ORDER, { orderId: params.orderId! }),
         label: getRouteLabel(LearnerDashboardPaths.ORDER_RUNS),
       },
     ],
