+++ src/frontend/js/widgets/Dashboard/hooks/useDashboardRouter/index.tsx
@@ -1,14 +1,9 @@
-import type {
-  FormatXMLElementFn,
-  Options as IntlMessageFormatOptions,
-  PrimitiveType,
-} from 'intl-messageformat';
 import { useMemo } from 'react';
 import { MessageDescriptor, useIntl } from 'react-intl';
-import { createBrowserRouter, NavigateOptions, useNavigate } from 'react-router-dom';
-import { LearnerDashboardPaths } from 'widgets/Dashboard/utils/learnerRouteMessages';
-import { getDashboardRoutes, getDashboardRoutePath } from 'widgets/Dashboard/utils/dashboardRoutes';
-import { TeacherDashboardPaths } from 'widgets/Dashboard/utils/teacherRouteMessages';
+import { createBrowserRouter, generatePath, NavigateOptions, useNavigate } from 'react-router-dom';
+import { getDashboardRoutes } from 'widgets/Dashboard/utils/dashboardRoutes';
+import { TeacherDashboardPaths } from 'widgets/Dashboard/utils/teacherDashboardPaths';
+import { LearnerDashboardPaths } from 'widgets/Dashboard/utils/learnerRoutesPaths';
 import { getDashboardBasename } from './getDashboardBasename';
 
 export interface DashboardRouteHandle {
@@ -31,20 +26,19 @@ const useDashboardRouter = () => {
  * Wrapper for `useNavigate` to avoid repetitive hooks calls.
  */
 export const useDashboardNavigate = () => {
-  const getRoutePath = getDashboardRoutePath(useIntl());
   const navigate = useNavigate();
+
   return useMemo(
     () =>
-      (
-        to: number | LearnerDashboardPaths | TeacherDashboardPaths,
-        values?: Record<string, PrimitiveType | FormatXMLElementFn<string, string>>,
-        options?: IntlMessageFormatOptions,
+      <Path extends LearnerDashboardPaths | TeacherDashboardPaths>(
+        to: number | Path,
+        params?: Parameters<typeof generatePath<Path>>[1],
         routerOptions?: NavigateOptions,
       ) => {
         if (typeof to === 'number') {
           return navigate(to);
         }
-        return navigate(getRoutePath(to, values, options), routerOptions);
+        return navigate(generatePath(to, params), routerOptions);
       },
     [],
   );
