+++ src/frontend/js/widgets/Dashboard/utils/learnerRoutes.tsx
@@ -7,17 +7,14 @@ import { DashboardPreferences } from 'pages/DashboardPreferences';
 import { DashboardEditCreditCardLoader } from 'pages/DashboardCreditCardsManagement/DashboardEditCreditCardLoader';
 import { DashboardCourses } from 'pages/DashboardCourses';
 import { DashboardOrderLoader } from 'widgets/Dashboard/components/DashboardOrderLoader';
-import {
-  getDashboardRouteLabel,
-  getDashboardRoutePath,
-} from 'widgets/Dashboard/utils/dashboardRoutes';
-import {
-  LEARNER_DASHBOARD_ROUTE_LABELS,
-  LearnerDashboardPaths,
-} from 'widgets/Dashboard/utils/learnerRouteMessages';
+import { getDashboardRouteLabel } from 'widgets/Dashboard/utils/dashboardRoutes';
 import { DashboardCertificates } from 'pages/DashboardCertificates';
 import { DashboardOrderLayout } from 'pages/DashboardOrderLayout';
 import { DashboardContracts } from 'pages/DashboardContracts';
+import {
+  LEARNER_DASHBOARD_ROUTE_LABELS,
+  LearnerDashboardPaths,
+} from 'widgets/Dashboard/utils/learnerRoutesPaths';
 
 export interface DashboardRouteHandle {
   crumbLabel?: MessageDescriptor;
@@ -26,11 +23,10 @@ export interface DashboardRouteHandle {
 
 export function getLearnerDashboardRoutes() {
   const intl = useIntl();
-  const getRoutePath = getDashboardRoutePath(intl);
   const getRouteLabel = getDashboardRouteLabel(intl);
   const routes: RouteObject[] = [
     {
-      path: getRoutePath(LearnerDashboardPaths.COURSES),
+      path: LearnerDashboardPaths.COURSES,
       handle: {
         crumbLabel: LEARNER_DASHBOARD_ROUTE_LABELS[LearnerDashboardPaths.COURSES],
       },
@@ -41,9 +37,7 @@ export function getLearnerDashboardRoutes() {
           element: <DashboardCourses />,
         },
         {
-          path: getRoutePath(LearnerDashboardPaths.ORDER, {
-            orderId: ':orderId',
-          }),
+          path: LearnerDashboardPaths.ORDER,
           element: <DashboardOrderLayout />,
           handle: {
             renderLayout: true,
@@ -59,22 +53,22 @@ export function getLearnerDashboardRoutes() {
       ],
     },
     {
-      path: getRoutePath(LearnerDashboardPaths.CERTIFICATES),
+      path: LearnerDashboardPaths.CERTIFICATES,
       handle: { crumbLabel: LEARNER_DASHBOARD_ROUTE_LABELS[LearnerDashboardPaths.CERTIFICATES] },
       element: <DashboardCertificates />,
     },
     {
-      path: getRoutePath(LearnerDashboardPaths.CONTRACTS),
+      path: LearnerDashboardPaths.CONTRACTS,
       handle: { crumbLabel: LEARNER_DASHBOARD_ROUTE_LABELS[LearnerDashboardPaths.CONTRACTS] },
       element: <DashboardContracts />,
     },
     {
-      path: getRoutePath(LearnerDashboardPaths.COURSE, { code: ':code' }),
+      path: LearnerDashboardPaths.COURSE,
       element: <RouteInfo title={getRouteLabel(LearnerDashboardPaths.COURSE)} />,
       handle: { crumbLabel: LEARNER_DASHBOARD_ROUTE_LABELS[LearnerDashboardPaths.COURSE] },
     },
     {
-      path: getRoutePath(LearnerDashboardPaths.PREFERENCES),
+      path: LearnerDashboardPaths.PREFERENCES,
       handle: { crumbLabel: LEARNER_DASHBOARD_ROUTE_LABELS[LearnerDashboardPaths.PREFERENCES] },
       element: <Outlet />,
       children: [
@@ -83,9 +77,7 @@ export function getLearnerDashboardRoutes() {
           element: <DashboardPreferences />,
         },
         {
-          path: getRoutePath(LearnerDashboardPaths.PREFERENCES_ADDRESS_EDITION, {
-            addressId: ':addressId',
-          }),
+          path: LearnerDashboardPaths.PREFERENCES_ADDRESS_EDITION,
           element: <DashboardEditAddressLoader />,
           handle: {
             crumbLabel:
@@ -93,7 +85,7 @@ export function getLearnerDashboardRoutes() {
           },
         },
         {
-          path: getRoutePath(LearnerDashboardPaths.PREFERENCES_ADDRESS_CREATION),
+          path: LearnerDashboardPaths.PREFERENCES_ADDRESS_CREATION,
           element: <DashboardCreateAddressLoader />,
           handle: {
             crumbLabel:
@@ -101,9 +93,7 @@ export function getLearnerDashboardRoutes() {
           },
         },
         {
-          path: getRoutePath(LearnerDashboardPaths.PREFERENCES_CREDIT_CARD_EDITION, {
-            creditCardId: ':creditCardId',
-          }),
+          path: LearnerDashboardPaths.PREFERENCES_CREDIT_CARD_EDITION,
           element: <DashboardEditCreditCardLoader />,
           handle: {
             crumbLabel:
